Separate Linux EGL integration from transparent window handling
