+++ Source/visualiser/OpenGL_linux.h
@@ -1,0 +1,7 @@
+#pragma once
+
+struct _XDisplay;
+
+namespace osci {
+bool hasCompatibleOpenGLVisual(_XDisplay* display);
+}
